changes in seed for security

+++ src/seed/seed.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { AuthService } from 'src/auth/auth.service';
@@ -10,22 +11,31 @@ import { initialData } from './data/seed-data';
 export class SeedService {
   constructor(
     private readonly propertiesService: PropertiesService,
-    private readonly authService: AuthService
+    private readonly authService: AuthService,
+    private readonly configService: ConfigService
   ) {}
   
   async runSeed() {
-    await this.propertiesService.removeAll();
-    await this.authService.removeAll();
-    
-    const seedUsers = initialData.users;
-    seedUsers.forEach(user => {
-      this.authService.create(user)
-    })
+    if (this.configService.get('NODE_ENV') === 'development') {
 
-    const seedProperties = initialData.properties;
-    seedProperties.forEach(property => {
-      this.propertiesService.create(property)
-    })
-    return 'SEED EXECUTED';
+      await this.propertiesService.removeAll();
+      await this.authService.removeAll();
+      
+      const seedUsers = initialData.users;
+      seedUsers.forEach(user => {
+        this.authService.create(user)
+      })
+  
+      const seedProperties = initialData.properties;
+      seedProperties.forEach(property => {
+        this.propertiesService.create(property)
+      })
+      return 'SEED EXECUTED';
+
+    } 
+    else {
+      return 'FORBIDDEN';
+    }
+
   }
 }
